fix hashmap
