+++ pkg/container/hashtable/int64_hash_map_2.go
@@ -20,31 +20,34 @@ import (
 	"unsafe"
 )
 
-func (ht *Int64HashMap) Free2(m *mpool.MPool) {
+func (ht *Int64HashMap) Free(m *mpool.MPool) {
 	for i := range ht.rawData2 {
-		m.Free(ht.rawData2[i])
+		if len(ht.rawData2[i]) > 0 {
+			m.Free(ht.rawData2[i])
+		}
 		ht.rawData2[i], ht.cells2[i] = nil, nil
 	}
 	ht.rawData2, ht.cells2 = nil, nil
 }
 
-func (ht *Int64HashMap) Init2(m *mpool.MPool) (err error) {
-	ht.cellCntBits = kInitialCellCntBits
+func (ht *Int64HashMap) Init(m *mpool.MPool) (err error) {
+	ht.blockCellCntBits = kInitialCellCntBits
 	ht.blockMaxCellCnt = kInitialCellCnt
-	ht.cellCntMask = kInitialCellCnt - 1
-	ht.allElemCnt = 0
 	ht.blockMaxElemCnt = kInitialCellCnt * kLoadFactorNumerator / kLoadFactorDenominator
+	ht.cellCntMask = kInitialCellCnt - 1
+	ht.elemCnt = 0
+	ht.allCellCnt = kInitialCellCnt
 
 	ht.rawData2 = make([][]byte, 1)
 	ht.cells2 = make([][]Int64HashMapCell, 1)
 	if ht.rawData2[0], err = m.Alloc(int(ht.blockMaxCellCnt) * int(intCellSize)); err == nil {
-		ht.cells2[0] = unsafe.Slice((*Int64HashMapCell)(unsafe.Pointer(&ht.rawData2[0])), ht.cellCnt)
+		ht.cells2[0] = unsafe.Slice((*Int64HashMapCell)(unsafe.Pointer(&ht.rawData2[0])), ht.blockMaxCellCnt)
 	}
 	return
 }
 
-func (ht *Int64HashMap) InsertBatch2(n int, hashes []uint64, keysPtr unsafe.Pointer, values []uint64, m *mpool.MPool) error {
-	if err := ht.resizeOnDemand2(n, m); err != nil {
+func (ht *Int64HashMap) InsertBatch(n int, hashes []uint64, keysPtr unsafe.Pointer, values []uint64, m *mpool.MPool) error {
+	if err := ht.resizeOnDemand(n, m); err != nil {
 		return err
 	}
 
@@ -54,19 +57,19 @@ func (ht *Int64HashMap) InsertBatch2(n int, hashes []uint64, keysPtr unsafe.Poin
 
 	keys := unsafe.Slice((*uint64)(keysPtr), n)
 	for i, key := range keys {
-		cell := ht.findCell2(hashes[i], key)
+		cell := ht.findCell(hashes[i], key)
 		if cell.Mapped == 0 {
-			ht.allElemCnt++
+			ht.elemCnt++
 			cell.Key = key
-			cell.Mapped = ht.allElemCnt
+			cell.Mapped = ht.elemCnt
 		}
 		values[i] = cell.Mapped
 	}
 	return nil
 }
 
-func (ht *Int64HashMap) InsertBatchWithRing2(n int, zValues []int64, hashes []uint64, keysPtr unsafe.Pointer, values []uint64, m *mpool.MPool) error {
-	if err := ht.resizeOnDemand2(n, m); err != nil {
+func (ht *Int64HashMap) InsertBatchWithRing(n int, zValues []int64, hashes []uint64, keysPtr unsafe.Pointer, values []uint64, m *mpool.MPool) error {
+	if err := ht.resizeOnDemand(n, m); err != nil {
 		return err
 	}
 
@@ -79,30 +82,30 @@ func (ht *Int64HashMap) InsertBatchWithRing2(n int, zValues []int64, hashes []ui
 		if zValues[i] == 0 {
 			continue
 		}
-		cell := ht.findCell2(hashes[i], key)
+		cell := ht.findCell(hashes[i], key)
 		if cell.Mapped == 0 {
-			ht.allElemCnt++
+			ht.elemCnt++
 			cell.Key = key
-			cell.Mapped = ht.allElemCnt
+			cell.Mapped = ht.elemCnt
 		}
 		values[i] = cell.Mapped
 	}
 	return nil
 }
 
-func (ht *Int64HashMap) FindBatch2(n int, hashes []uint64, keysPtr unsafe.Pointer, values []uint64) {
+func (ht *Int64HashMap) FindBatch(n int, hashes []uint64, keysPtr unsafe.Pointer, values []uint64) {
 	if hashes[0] == 0 {
 		Int64BatchHash(keysPtr, &hashes[0], n)
 	}
 
 	keys := unsafe.Slice((*uint64)(keysPtr), n)
 	for i, key := range keys {
-		cell := ht.findCell2(hashes[i], key)
+		cell := ht.findCell(hashes[i], key)
 		values[i] = cell.Mapped
 	}
 }
 
-func (ht *Int64HashMap) FindBatchWithRing2(n int, zValues []int64, hashes []uint64, keysPtr unsafe.Pointer, values []uint64) {
+func (ht *Int64HashMap) FindBatchWithRing(n int, zValues []int64, hashes []uint64, keysPtr unsafe.Pointer, values []uint64) {
 	if hashes[0] == 0 {
 		Int64BatchHash(keysPtr, &hashes[0], n)
 	}
@@ -113,44 +116,64 @@ func (ht *Int64HashMap) FindBatchWithRing2(n int, zValues []int64, hashes []uint
 			values[i] = 0
 			continue
 		}
-		cell := ht.findCell2(hashes[i], key)
+		cell := ht.findCell(hashes[i], key)
 		values[i] = cell.Mapped
 	}
 }
 
-func (ht *Int64HashMap) findCell2(hash uint64, key uint64) *Int64HashMapCell {
-	for idx := hash & ht.cellCntMask; true; idx = (idx + 1) & ht.cellCntMask {
-		blockId := idx / ht.blockMaxCellCnt
-		cellId := idx % ht.blockMaxCellCnt
-		cell := &ht.cells2[blockId][cellId]
-		if cell.Key == key || cell.Mapped == 0 {
-			return cell
+func (ht *Int64HashMap) findCell(hash uint64, key uint64) *Int64HashMapCell {
+	if len(ht.rawData2) == 1 {
+		for idx := hash & ht.cellCntMask; true; idx = (idx + 1) & ht.cellCntMask {
+			cell := &ht.cells2[0][idx]
+			if cell.Key == key || cell.Mapped == 0 {
+				return cell
+			}
+		}
+	} else {
+		for idx := hash & ht.cellCntMask; true; idx = (idx + 1) & ht.cellCntMask {
+			blockId := idx / ht.blockMaxCellCnt
+			cellId := idx % ht.blockMaxCellCnt
+			cell := &ht.cells2[blockId][cellId]
+			if cell.Key == key || cell.Mapped == 0 {
+				return cell
+			}
 		}
 	}
+
 	return nil
 }
 
-func (ht *Int64HashMap) findEmptyCell2(hash uint64, key uint64) *Int64HashMapCell {
-	for idx := hash & ht.cellCntMask; true; idx = (idx + 1) & ht.cellCntMask {
-		blockId := idx / ht.blockMaxCellCnt
-		cellId := idx % ht.blockMaxCellCnt
-		cell := &ht.cells2[blockId][cellId]
-		if cell.Mapped == 0 {
-			return cell
+func (ht *Int64HashMap) findEmptyCell(hash uint64, key uint64) *Int64HashMapCell {
+	if len(ht.rawData2) == 1 {
+		for idx := hash & ht.cellCntMask; true; idx = (idx + 1) & ht.cellCntMask {
+			cell := &ht.cells2[0][idx]
+			if cell.Mapped == 0 {
+				return cell
+			}
+		}
+	} else {
+		for idx := hash & ht.cellCntMask; true; idx = (idx + 1) & ht.cellCntMask {
+			blockId := idx / ht.blockMaxCellCnt
+			cellId := idx % ht.blockMaxCellCnt
+			cell := &ht.cells2[blockId][cellId]
+			if cell.Mapped == 0 {
+				return cell
+			}
 		}
 	}
+
 	return nil
 }
 
-func (ht *Int64HashMap) resizeOnDemand2(n int, m *mpool.MPool) error {
-	targetCnt := ht.allElemCnt + uint64(n)
+func (ht *Int64HashMap) resizeOnDemand(n int, m *mpool.MPool) error {
+	targetCnt := ht.elemCnt + uint64(n)
 	if targetCnt < uint64(len(ht.rawData2))*ht.blockMaxElemCnt {
 		return nil
 	}
 
 	var err error
 	if len(ht.rawData2) == 1 {
-		newCellCntBits := ht.cellCntBits + 2
+		newCellCntBits := ht.blockCellCntBits + 2
 		newCellCnt := uint64(1 << newCellCntBits)
 		newBlockMaxElemCnt := newCellCnt * kLoadFactorNumerator / kLoadFactorDenominator
 		for newBlockMaxElemCnt < targetCnt {
@@ -166,16 +189,17 @@ func (ht *Int64HashMap) resizeOnDemand2(n int, m *mpool.MPool) error {
 		newAlloc := int(newCellCnt) * int(intCellSize)
 		if newAlloc < mpool.GB {
 			// update hashTable cnt.
-			ht.cellCntBits = newCellCntBits
+			ht.blockCellCntBits = newCellCntBits
 			ht.blockMaxCellCnt = newCellCnt
-			ht.cellCntMask = newCellCnt - 1
 			ht.blockMaxElemCnt = newBlockMaxElemCnt
+			ht.allCellCnt = newCellCnt
+			ht.cellCntMask = newCellCnt - 1
 
 			ht.rawData2[0], err = m.Alloc(newAlloc)
 			if err != nil {
 				return err
 			}
-			ht.cells2[0] = unsafe.Slice((*Int64HashMapCell)(unsafe.Pointer(&ht.rawData2[0][0])), ht.cellCnt)
+			ht.cells2[0] = unsafe.Slice((*Int64HashMapCell)(unsafe.Pointer(&ht.rawData2[0][0])), ht.blockMaxCellCnt)
 
 			// re the cell
 			var hashes [256]uint64
@@ -185,7 +209,7 @@ func (ht *Int64HashMap) resizeOnDemand2(n int, m *mpool.MPool) error {
 				for j := range cells {
 					cell := &cells[j]
 					if cell.Mapped != 0 {
-						newCell := ht.findEmptyCell2(hashes[j], cell.Key)
+						newCell := ht.findEmptyCell(hashes[j], cell.Key)
 						*newCell = *cell
 					}
 				}
@@ -207,15 +231,17 @@ func (ht *Int64HashMap) resizeOnDemand2(n int, m *mpool.MPool) error {
 		}
 		ht.cells2[newBlockIdx] = unsafe.Slice((*Int64HashMapCell)(unsafe.Pointer(&ht.rawData2[newBlockIdx][0])), ht.blockMaxCellCnt)
 	}
+	ht.allCellCnt = ht.blockMaxCellCnt * uint64(len(ht.rawData2))
+	ht.cellCntMask = ht.allCellCnt - 1
 	return nil
 }
 
-func (ht *Int64HashMap) Cardinality2() uint64 {
-	return ht.allElemCnt
+func (ht *Int64HashMap) Cardinality() uint64 {
+	return ht.elemCnt
 }
 
-func (it *Int64HashMapIterator) Next2() (cell *Int64HashMapCell, err error) {
-	for it.pos < it.table.cellCnt {
+func (it *Int64HashMapIterator) Next() (cell *Int64HashMapCell, err error) {
+	for it.pos < it.table.allCellCnt {
 		blockId := it.pos / it.table.blockMaxCellCnt
 		cellId := it.pos % it.table.blockMaxCellCnt
 		cell = &it.table.cells2[blockId][cellId]
